Text Data Pre-processing Using Vector Space Model
Big Data is a term which is used to describe massive amount of data generating from digital sources or the internet. From the past few years data is exponentially growing due to the use of connected devices such as smart phone’s, tablets, laptops and desktop computer. Moreover E-commerce which is also known as online market, internet services and social networking sites are generating tremendous user data in the form of documents, emails and web pages. This generated data volume is so vast and overwhelming which makes complex to process and analyze using traditional software systems consuming more time. This project presents a pre-processing algorithm to extract real time user accessed data from windows operating system environment.
Text Pre-processing is an important task and critical step in Text mining, Natural Language Processing and information retrieval. In the area of Text Mining, data pre-processing used for extracting interesting and non-trivial and knowledge from unstructured text data. Information Retrieval is essentially a matter of deciding which documents in a collection should be retrieved to satisfy a user's need for information. The user's need for information is represented by a query or profile, and contains one or more search terms, plus some additional information such as weight of the words. Hence, the retrieval decision is made by comparing the terms of the query with the index appearing in the document itself. The decision may be binary (retrieve/reject), or it may involve estimating the degree of relevance that the document has to query. Unfortunately, the words that appear in documents and in queries often have many structural variants. So, before the information retrieval from the documents, the data pre-processing techniques are applied on the target data set to reduce the size of the data set which will increase the effectiveness of IR System The objective of this project is to analyse the issues of pre-processing methods such as Tokenization, Stop word removal, Term weighting and to obtain the result in shot time